+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.5.0
+
+* exported new server struct
+* exposed server router and content
+* fixed concurrency issue
+
 ## 0.4.0
 
 * fixed concurrency issue